♻️ refactor: api 파일 js 리팩토링(에러 핸들링 위치 변경)
