--- week4_login/src/apis/join/memberJoin.jsx
@@ -1,11 +1,0 @@
-import axios from 'axios';
-
-// 회원가입 API 함수
-export const memberJoin = async (requestBody) => {
-  const response = await axios.post(
-    `${import.meta.env.VITE_BASE_URL}/member/join`,
-    requestBody
-  );
-
-  return response;
-};
